#zapcot-85 - Teste para ordenação da CTEEP
